+++ fengoffice_/public/assets/javascript/ckeditor/ck_upload_handler.php
@@ -22,6 +22,12 @@
 		$file_name = preg_replace("/[^a-zA-Z0-9.]/", "_", $file_name);
 		$file_url = ROOT_URL . "/tmp/" . $file_name;
 		
+		//Exclude non image files.
+       /* $pattern = '/^.*\.(jpeg|JPEG|jpg|JPG|gif|GIF|png|PNG|bmp|BMP)$/';
+        preg_match($pattern, $file_name, $matches, PREG_OFFSET_CAPTURE);
+        if(count($matches) <= 0)
+        	die('Invalid File');*/
+        	
 		copy($file_info['tmp_name'], TEMP_PATH . DIRECTORY_SEPARATOR . $file_name);
 		unlink($file_info['tmp_name']);
